Convert to Docker buildx
With the new [https://docs.docker.com/docker-for-mac/multi-arch/](multi-arch)
support in Docker, it's no longer necessary to build manifests manually
and push them.

This commit removes the separate Dockerfiles for each architecture
in favor of a single Dockerfile using `python:alpine` as the base
image. It also removes the `build.sh` script and replaces it with a
simple `Makefile`.

The `Makefile` has a target to build a local image for development
as well as a target for a multi-arch build with `buildx` that pushes
to DockerHub.